GVT-1733: Move code common to element and vertical geometry listings to ListingCommon
